% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_list.R
\name{clean_list}
\alias{clean_list}
\title{Clean and simplify a list}
\usage{
clean_list(lst, format = FALSE)
}
\arguments{
\item{lst}{A list object.}
}
\value{
Returns a simplified list. It is possible that \code{lst} will be returned
unmodified if no lame lists or leaf tables are found.
}
\description{
Cleans a list by pruning (removing) unnecessary nested lists and by simplifying
leaf nodes into dataframes when appropriate.
}
\examples{
clean_list(list(A = 1:10, B = 11:20))

clean_list(list(A = list(X = 1:10, Y = letters[1:10])))

clean_list(list(A = 1:10, B = "TEXT", C = c("FOO", "BAR")))

}
