% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_list.R
\name{collaspe_leaf}
\alias{collaspe_leaf}
\title{Convert vector to a scalar}
\usage{
collaspe_leaf(x)
}
\arguments{
\item{x}{A vector or object in a list.}
}
\value{
Returns a character vector of length 1.
}
\description{
Collapses a vector to a scalar.
}
\details{
Converting a vector to a scalar is useful for converting a list in R to an
html list.
}
