% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_list.R
\name{collaspe_leaf_r}
\alias{collaspe_leaf_r}
\title{Recursively apply \code{collapse_leaf}}
\usage{
collaspe_leaf_r(lst)
}
\arguments{
\item{lst}{A list object.}
}
\value{
Returns a list.
}
\description{
Recursively applies \code{collapse_leaf} to a list object.
}
