% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_list.R
\name{lame_list}
\alias{lame_list}
\title{Detect an unnecessary neste list}
\usage{
lame_list(lst)
}
\arguments{
\item{lst}{A list object}
}
\value{
Returns \code{TRUE} if list object is a lame list, \code{FALSE} if not.
}
\description{
If a list object contains an unnamed list containing another list or a single
object it is a "lame list". For example: \code{list(list(1:10))}. The nested
list in this in this example is unnecessary since it contains a single vector,
therefore can be removed. Pruning lame list can help simplify a messy list.
}
\examples{
lame_list(list(list(1:10)))

lame_list(list(A = list(1:10)))

lame_list(list(list(1:10, list(11:20))))
}
\seealso{
\code{\link{clean_list}}
}
