% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_list.R
\name{list_depth}
\alias{list_depth}
\title{Find list depth}
\usage{
list_depth(lst)
}
\arguments{
\item{lst}{A list object.}
}
\value{
Returns a numeric vetor of list depths.
}
\description{
Iterates \code{\link{purrr::vec_depth}} over each object in a list.
}
