% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_html.R
\name{list_html}
\alias{list_html}
\title{Create an HTML list item}
\usage{
list_html(lst)
}
\arguments{
\item{lst}{A list object.}
}
\value{
Returns an HTML string of a list item.
}
\description{
Converts a list object to an HTML list item. If an object is a
dataframe or matrix it is convert to an HTML table first.
}
