% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_html.R
\name{list_html_r}
\alias{list_html_r}
\title{Create an un-ordered HTML list recursively}
\usage{
list_html_r(lst)
}
\arguments{
\item{lst}{A list object.}
}
\value{
Returns a list of HTML list item strings.
}
\description{
Applies \code{\link{list_html}} to every entry of a list object.
}
