% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_list.R
\name{list_length}
\alias{list_length}
\title{Find list length}
\usage{
list_length(lst)
}
\arguments{
\item{lst}{A list object.}
}
\value{
Returns a numeric vector of list lengths.
}
\description{
Iterates \code{length} over each object in a list
}
