% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_list.R
\name{list_to_df}
\alias{list_to_df}
\title{Convert leaf node to a dataframe}
\usage{
list_to_df(lst)
}
\arguments{
\item{lst}{A list object.}
}
\description{
Determines whether a leaf node contains an object that can be converted to a
dataframe.
}
\seealso{
\code{\link{is_leaf_table}}, \code{\link{simplify_list}}
}
