% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_html.R
\name{list_to_html}
\alias{list_to_html}
\title{Convert list to HTML}
\usage{
list_to_html(lst)
}
\arguments{
\item{lst}{A list object.}
}
\value{
Returns an un-ordered HTML list.
}
\description{
Converts a list to an un-ordered HTML list.
}
