% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_list.R
\name{simplify_list}
\alias{simplify_list}
\title{Remove unnecessary nested lists}
\usage{
simplify_list(lst)
}
\arguments{
\item{lst}{A list object}
}
\description{
Removes unnecessary lists (or "lame lists") from a list. A lame list is a
unnamed list contain a single list object(see \code{\link{lame_list}}).
}
\seealso{
\code{\link{simplify_list}} \code{\link{lame_list}}
}
